no message
svn path=/trunk/yarp2/; revision=4717

+++ src/libYARP_dev/src/dragonfly/linux/yarp/DragonflyDeviceDriver.cpp
@@ -119,12 +119,12 @@ inline bool DragonflyResources::_initialize (const DragonflyOpenParameters& para
 	if (params._whiteR>0) {
 		printf("white balance %g %g\n", 
                (double)params._whiteR, (double)params._whiteB);
-		cam.SetColor(params._whiteR,params._whiteB); 
+		cam.SetColor((float)params._whiteR,(float)params._whiteB); 
 	} else {
 		cam.SetColor((float)50.0/63, (float)20.0/63); 
     }
-	cam.SetShutter(params._shutter);	// x * 0.0625 = 20 mSec = 50 Hz
-	cam.SetGain(params._gain);		// x * -0.0224 = -11.2dB
+	cam.SetShutter((float)params._shutter);	// x * 0.0625 = 20 mSec = 50 Hz
+	cam.SetGain((float)params._gain);		// x * -0.0224 = -11.2dB
     //    cam.SetBrigthness(params._brightness);
     //    cam.SetExposure(params._exposure);
